Restructure arm_controller as library and create test_arm_controller as a test executable
