+++ lib/include/arm_controller/arm_controller.h
@@ -187,4 +187,3 @@ public:
     void moveToRest();
 
 };
-
